chore: delete DashboardDefaultWidgetTab
