+++ src/services/dashboards/dashboard-customize/modules/DashboardAddWidgetModal.vue
@@ -9,7 +9,48 @@
                    :active-tab.sync="tabState.activeTab"
             >
                 <template #default-widget>
-                    <dashboard-default-widget-tab />
+                    <div class="dashboard-default-widget-tab">
+                        <div class="left-area">
+                            <p-button-tab :tabs="subTabState.tabs"
+                                          :active-tab="subTabState.activeTab"
+                                          @change="handleChangeTab"
+                            />
+                            <ul class="widget-list">
+                                <li v-for="widget in widgets"
+                                    :key="`card-${widget.widget_config_id}`"
+                                    class="widget-card"
+                                    :class="{'selected' : selectedWidgetConfigId === widget.widget_config_id}"
+                                    @click="selectWidget(widget.widget_config_id)"
+                                >
+                                    <div class="card-header">
+                                        {{ widget.title }}
+                                    </div>
+                                    <div class="card-content">
+                                        <p-lazy-img :src="`/images/widgets/${widget.description?.preview_image}`"
+                                                    class="preview-img"
+                                                    height="4.75rem"
+                                                    width="100%"
+                                        >
+                                            <template #error>
+                                                <img src="/images/widgets/widget-img_default--thumbnail.png">
+                                            </template>
+                                        </p-lazy-img>
+                                    </div>
+                                </li>
+                            </ul>
+                        </div>
+                        <div class="right-area">
+                            <div v-if="!selectedWidgetConfigId"
+                                 class="no-selected-wrapper"
+                            >
+                                <span class="title">{{ $t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.NO_SELECTED') }}</span>
+                                <span class="text">{{ $t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.NO_SELECTED_HELP_TEXT') }}</span>
+                            </div>
+                            <dashboard-add-default-widget-right-area v-else
+                                                                     :widget-config-id="selectedWidgetConfigId"
+                            />
+                        </div>
+                    </div>
                 </template>
             </p-tab>
         </template>
@@ -21,14 +62,19 @@ import {
     computed, defineComponent, reactive, toRefs,
 } from 'vue';
 
-import { PButtonModal, PTab } from '@spaceone/design-system';
+import {
+    PButtonModal, PTab, PButtonTab, PLazyImg,
+} from '@spaceone/design-system';
 import type { TabItem } from '@spaceone/design-system/src/navigation/tabs/tab/type';
 
 import { i18n } from '@/translations';
 
 import { useProxyValue } from '@/common/composables/proxy-state';
 
-import DashboardDefaultWidgetTab from '@/services/dashboards/dashboard-customize/modules/DashboardDefaultWidgetTab.vue';
+import DashboardAddDefaultWidgetRightArea
+    from '@/services/dashboards/dashboard-customize/modules/DashboardAddDefaultWidgetRightArea.vue';
+import type { WidgetConfig } from '@/services/dashboards/widgets/config';
+import { CONSOLE_WIDGET_CONFIGS } from '@/services/dashboards/widgets/widget-config-list';
 
 interface Props {
     visible: boolean;
@@ -36,9 +82,11 @@ interface Props {
 export default defineComponent<Props>({
     name: 'DashboardAddWidgetModal',
     components: {
-        DashboardDefaultWidgetTab,
         PTab,
         PButtonModal,
+        PButtonTab,
+        PLazyImg,
+        DashboardAddDefaultWidgetRightArea,
     },
     props: {
         visible: {
@@ -49,6 +97,8 @@ export default defineComponent<Props>({
     setup(props, { emit }: SetupContext) {
         const state = reactive({
             proxyVisible: useProxyValue('visible', props, emit),
+            widgets: computed(() => getWidgetConfigsByLabel(subTabState.activeTab)),
+            selectedWidgetConfigId: undefined as string | undefined,
         });
         const tabState = reactive({
             tabs: computed<TabItem[]>(() => ([
@@ -57,10 +107,37 @@ export default defineComponent<Props>({
             ])),
             activeTab: 'default-widget',
         });
+        const subTabState = reactive({
+            tabs: computed<TabItem[]>(() => [
+                { name: 'all', label: `${i18n.t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.TAB_ALL')} (${getWidgetConfigsByLabel('all').length})` },
+                { name: 'Cost', label: `${i18n.t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.TAB_COST')} (${getWidgetConfigsByLabel('Cost').length})` },
+                { name: 'Asset', label: `${i18n.t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.TAB_ASSET')} (${getWidgetConfigsByLabel('Asset').length})` },
+                { name: 'Alert', label: `${i18n.t('DASHBOARDS.CUSTOMIZE.ADD_WIDGET.TAB_ALERT')} (${getWidgetConfigsByLabel('Alert').length})` },
+            ]),
+            activeTab: 'all',
+        });
+
+        /* Util */
+        const getWidgetConfigsByLabel = (label: string): Array<Partial<WidgetConfig>> => {
+            const allConfigs = Object.values(CONSOLE_WIDGET_CONFIGS);
+            if (label === 'all') return allConfigs;
+            return allConfigs.filter((config) => config.labels?.includes(label));
+        };
+
+        /* Event */
+        const handleChangeTab = (selectedTab) => {
+            tabState.activeTab = selectedTab;
+        };
+        const selectWidget = (widgetConfigId: string) => {
+            state.selectedWidgetConfigId = widgetConfigId;
+        };
 
         return {
             ...toRefs(state),
             tabState,
+            subTabState,
+            handleChangeTab,
+            selectWidget,
         };
     },
 });
@@ -69,6 +146,86 @@ export default defineComponent<Props>({
 .dashboard-add-widget-modal {
     .p-tab {
         @apply border-0;
+    }
+    .dashboard-default-widget-tab {
+        display: flex;
+        width: 100%;
+        height: inherit;
+
+        /* custom design-system component - p-button-tab */
+        :deep(.p-button-tab) {
+            .button-group {
+                margin-left: 0;
+                button:first-child {
+                    margin-left: 0;
+                }
+            }
+            .tab-pane {
+                display: none;
+            }
+        }
+        .left-area {
+            @apply flex flex-col flex-grow flex-shrink border-r border-gray-200;
+            .widget-list {
+                @apply grid grid-cols-3 gap-2;
+                height: 100%;
+                overflow: auto;
+                padding-right: 1.25rem;
+                .widget-card {
+                    @apply flex flex-col overflow-hidden border border-gray-300 rounded-md cursor-pointer;
+                    height: 9.125rem;
+
+                    &.selected {
+                        @apply border border-blue-600;
+                        .card-header {
+                            @apply text-blue-600;
+                        }
+                    }
+                    .card-header {
+                        @apply flex items-center flex-grow;
+                        font-size: 0.875rem;
+                        height: 3.125rem;
+                        padding: 1rem;
+                    }
+                    .card-content {
+                        @apply flex justify-center items-center bg-gray-100;
+                        height: 5.875rem;
+
+                        /* custom design-system component - p-lazy-img */
+                        :deep(.preview-img) {
+                            .img-container {
+                                display: flex;
+                                justify-content: center;
+                                width: 100%;
+
+                                img {
+                                    width: auto !important;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        .right-area {
+            @apply overflow-auto flex-shrink-0;
+            width: 20rem;
+            height: inherit;
+            padding: 1.25rem;
+            .no-selected-wrapper {
+                @apply flex flex-col justify-center items-center;
+                height: 100%;
+                font-size: 0.875rem;
+                text-align: center;
+                .title {
+                    @apply text-primary-2;
+                    padding-bottom: 0.75rem;
+                }
+                .text {
+                    @apply text-gray-600;
+                }
+            }
+        }
     }
     &.p-button-modal {
         $modal-hedaer: 3.5rem;
